app/requirements.txt
fastapi==0.100.1
uvicorn==0.23.2
sqlalchemy==2.0.19
sqladmin==0.14.0
fastapi_admin==1.0.4
passlib==1.7.4
pandas==2.0.3
numpy==1.25.2
bcrypt==4.0.1
uuid==1.30
python-jose==3.3.0
python-multipart==0.0.6
async-exit-stack==1.0.1
async-generator==1.10
gunicorn==21.2.0
psycopg2==2.9.7
pyjwt==2.8.0
# jwt==1.3.1s

